chateau_rouge/requirements.txt
aiodns==3.1.1
aiosmtpd==1.4.5
asgiref==3.7.2
atpublic==4.0
attrs==23.2.0
cffi==1.16.0
deprecation==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
django-email-verification==0.3.3
django-jazzmin==3.0.0
django-use-email-as-username==1.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
npm==0.1.1
optional-django==0.1.0
packaging==24.0
pillow==10.2.0
pycares==4.4.0
pycparser==2.21
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
validate-email==1.3
validators==0.22.0
verify-email==2.4.3
